requirements.txt
selenium
ipykernel
python-dotenv
pyyaml
openai